reviews

+++ backend/migrations/20210125183151-create-product.js
@@ -16,7 +16,7 @@ module.exports = {
         type: Sequelize.TEXT
       },
       price: {
-        type: Sequelize.NUMERIC,
+        type: Sequelize.FLOAT(5,2),
         allowNull: false,
       },
       imgPath: {
